load fuul project
made a pseudo multiplayer, where there are two cameras, each of which observes with its own prefab-player. Control is demarcated via hardcode to the name of the prefab. Not very good, but I did not think of another quick solution. The very movement of cars followed through physics, to the direct control of the transform